# openapi/payments.yaml
openapi: 3.0.3
info:
  title: Insforge Payments API
  version: 1.0.0
  description: Provider-specific Stripe and Razorpay payment setup, runtime flows, catalog management, sync, and webhook projections
...
paths:
...
  /api/payments/stripe/{environment}/webhook:
    post:
      summary: Configure Managed Stripe Webhook
      description: Create or recreate the InsForge-managed Stripe webhook endpoint for one environment.
      tags:
        - Stripe Payments
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentPath'
      responses:
        '200':
          description: Managed webhook configuration result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigureStripeWebhookResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/payments/razorpay/config:
    get:
      summary: Get Razorpay Key Configuration
      description: Return masked Razorpay key configuration for test and live environments.
      tags:
        - Razorpay Payments
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: Razorpay key configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRazorpayConfigResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/payments/razorpay/sync:
    post:
      summary: Sync Razorpay Payments State
      description: Sync items, plans, customers, subscriptions, invoices, and payments for every configured Razorpay environment. Razorpay remains the source of truth.
      tags:
        - Razorpay Payments
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: Sync result for configured Razorpay environments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncRazorpayPaymentsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/payments/razorpay/{environment}/config:
    put:
      summary: Configure Razorpay Keys
      description: Validate and store Razorpay API keys for one environment. A webhook secret is generated automatically when one does not already exist.
      tags:
        - Razorpay Payments
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertRazorpayConfigBody'
            example:
              keyId: "rzp_test_xxx"
              keySecret: "razorpay_secret_xxx"
              webhookSecret: "webhook_secret_xxx"
      responses:
        '200':
          description: Updated Razorpay key configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRazorpayConfigResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      summary: Remove Razorpay Keys
      description: Remove the configured Razorpay keys for one environment.
      tags:
        - Razorpay Payments
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentPath'
      responses:
        '200':
          description: Updated Razorpay key configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRazorpayConfigResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/payments/razorpay/{environment}/sync:
    post:
      summary: Sync Razorpay Payments State For One Environment
      description: Sync items, plans, customers, subscriptions, invoices, and payments for one Razorpay environment. Razorpay remains the source of truth.
      tags:
        - Razorpay Payments
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentPath'
      responses:
        '200':
          description: Sync result for the requested Razorpay environment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncRazorpayPaymentsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/payments/razorpay/{environment}/webhook:
    get:
      summary: Get Razorpay Webhook Setup Values
      description: Return the Razorpay webhook URL and signing secret that must be copied into the Razorpay Dashboard for manual webhook setup.
      tags:
        - Razorpay Payments
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentPath'
      responses:
        '200':
          description: Razorpay webhook setup values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRazorpayWebhookSetupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/payments/razorpay/{environment}/webhook/rotate-secret:
    post:
      summary: Rotate Razorpay Webhook Secret
      description: Generate a new Razorpay webhook signing secret. Update the webhook secret in the Razorpay Dashboard after calling this endpoint.
      tags:
        - Razorpay Payments
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/EnvironmentPath'
      responses:
        '200':
          description: Razorpay webhook setup values with the rotated secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateRazorpayWebhookSecretResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
    stripeSignature:
      type: apiKey
      in: header
      name: stripe-signature
    razorpaySignature:
      type: apiKey
      in: header
      name: x-razorpay-signature
  parameters:
    EnvironmentPath:
      name: environment
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PaymentEnvironment'
      description: Payment provider environment.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    PaymentEnvironment:
      type: string
      enum:
        - test
        - live
    StripeEnvironment:
      $ref: '#/components/schemas/PaymentEnvironment'
    RazorpayEnvironment:
      $ref: '#/components/schemas/PaymentEnvironment'
...
    StripeConnectionStatus:
      type: string
      enum:
        - unconfigured
        - connected
        - error
    StripeLatestSyncStatus:
      type: string
      enum:
        - succeeded
        - failed
    RazorpayConnectionStatus:
      type: string
      enum:
        - unconfigured
        - connected
        - error
    RazorpayLatestSyncStatus:
      type: string
      enum:
        - succeeded
        - failed
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - statusCode
      properties:
        error:
          type: string
          example: "INVALID_INPUT"
        message:
          type: string
          example: "environment: Invalid enum value"
        statusCode:
          type: integer
          example: 400
        nextActions:
          type: string
          nullable: true
    StripeConnection:
      type: object
      required:
        - environment
        - status
        - accountId
        - accountEmail
        - accountLivemode
        - webhookEndpointId
        - webhookEndpointUrl
        - webhookConfiguredAt
        - maskedKey
        - lastSyncedAt
        - lastSyncStatus
        - lastSyncError
        - lastSyncCounts
      properties:
        environment:
          $ref: '#/components/schemas/StripeEnvironment'
        status:
          $ref: '#/components/schemas/StripeConnectionStatus'
        accountId:
          type: string
          nullable: true
          example: "acct_123"
        accountEmail:
          type: string
          nullable: true
          example: "owner@example.com"
        accountLivemode:
          type: boolean
          nullable: true
        webhookEndpointId:
          type: string
          nullable: true
          example: "we_123"
        webhookEndpointUrl:
          type: string
          nullable: true
          example: "https://api.example.com/api/webhooks/stripe/test"
        webhookConfiguredAt:
          type: string
          format: date-time
          nullable: true
        maskedKey:
          type: string
          nullable: true
          example: "sk_test_...abcd"
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        lastSyncStatus:
          allOf:
            - $ref: '#/components/schemas/StripeLatestSyncStatus'
          nullable: true
        lastSyncError:
          type: string
          nullable: true
        lastSyncCounts:
          type: object
          additionalProperties:
            type: number
    RazorpayConnection:
      type: object
      required:
        - environment
        - status
        - accountId
        - merchantName
        - accountLivemode
        - webhookEndpointId
        - webhookEndpointUrl
        - webhookConfiguredAt
        - maskedKey
        - lastSyncedAt
        - lastSyncStatus
        - lastSyncError
        - lastSyncCounts
      properties:
        environment:
          $ref: '#/components/schemas/RazorpayEnvironment'
        status:
          $ref: '#/components/schemas/RazorpayConnectionStatus'
        accountId:
          type: string
          nullable: true
          example: "acc_123"
        merchantName:
          type: string
          nullable: true
          example: "Example Merchant"
        accountLivemode:
          type: boolean
          nullable: true
        webhookEndpointId:
          type: string
          nullable: true
          example: "manual"
        webhookEndpointUrl:
          type: string
          nullable: true
          example: "https://api.example.com/api/webhooks/razorpay/test"
        webhookConfiguredAt:
          type: string
          format: date-time
          nullable: true
        maskedKey:
          type: string
          nullable: true
          example: "rzp_test_...abcd"
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        lastSyncStatus:
          allOf:
            - $ref: '#/components/schemas/RazorpayLatestSyncStatus'
          nullable: true
        lastSyncError:
          type: string
          nullable: true
        lastSyncCounts:
          type: object
          additionalProperties:
            type: number
...
    RazorpayKeyConfig:
      type: object
      required:
        - environment
        - keyType
        - hasKey
        - maskedKey
      properties:
        environment:
          $ref: '#/components/schemas/RazorpayEnvironment'
        keyType:
          type: string
          enum:
            - api_key
            - api_secret
            - webhook_secret
        hasKey:
          type: boolean
        maskedKey:
          type: string
          nullable: true
          example: "rzp_test_...abcd"
...
    GetRazorpayConfigResponse:
      type: object
      required:
        - keys
      properties:
        keys:
          type: array
          items:
            $ref: '#/components/schemas/RazorpayKeyConfig'
...
    UpsertRazorpayConfigBody:
      type: object
      required:
        - keyId
        - keySecret
      additionalProperties: false
      properties:
        keyId:
          type: string
          minLength: 1
          example: "rzp_test_xxx"
        keySecret:
          type: string
          minLength: 1
          writeOnly: true
          example: "razorpay_secret_xxx"
        webhookSecret:
          type: string
          writeOnly: true
          example: "webhook_secret_xxx"
...
    RazorpaySyncCounts:
      type: object
      required:
        - plans
        - items
        - customers
        - subscriptions
        - invoices
        - payments
      additionalProperties: false
      properties:
        plans:
          type: integer
          minimum: 0
        items:
          type: integer
          minimum: 0
        customers:
          type: integer
          minimum: 0
        subscriptions:
          type: integer
          minimum: 0
        invoices:
          type: integer
          minimum: 0
        payments:
          type: integer
          minimum: 0
    SyncRazorpayPaymentsEnvironmentResult:
      type: object
      required:
        - environment
        - status
        - connection
        - syncCounts
        - error
      additionalProperties: false
      properties:
        environment:
          $ref: '#/components/schemas/RazorpayEnvironment'
        status:
          type: string
          enum:
            - succeeded
            - failed
        connection:
          $ref: '#/components/schemas/RazorpayConnection'
        syncCounts:
          $ref: '#/components/schemas/RazorpaySyncCounts'
        error:
          type: string
          nullable: true
    SyncRazorpayPaymentsResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SyncRazorpayPaymentsEnvironmentResult'
    ConfigureStripeWebhookResponse:
      type: object
      required:
        - connection
      properties:
        connection:
          $ref: '#/components/schemas/StripeConnection'
    GetRazorpayWebhookSetupResponse:
      type: object
      required:
        - connection
        - webhookUrl
        - webhookSecret
      properties:
        connection:
          $ref: '#/components/schemas/RazorpayConnection'
        webhookUrl:
          type: string
          description: Webhook endpoint URL to copy into the Razorpay Dashboard.
          example: "https://api.example.com/api/webhooks/razorpay/test"
        webhookSecret:
          type: string
          description: Raw signing secret to copy into the Razorpay Dashboard.
          example: "webhook_secret_xxx"
    RotateRazorpayWebhookSecretResponse:
      allOf:
        - $ref: '#/components/schemas/GetRazorpayWebhookSetupResponse'
...